Add files via upload
This code is used to set up a very simple server using Pythons' socket module on any device being used. In client/server protocol, the server must always being running before the client makes a request. Once this script is running, it will request the IP address of the device being used to setup as the server and send the message, "I am SERVER", to the client while the server waits for a message from the client which will assure an established connection.